Add supported firmware version check
Python driver has a list of purpledrop software version patterns that it
supports, and prints an error message if it detects anything else. I intend
to make non-backwards compatible changes to the message protocol, and I
don't intend to maintain support for older versions of the embedded software.
This will help users identify when there python and embedded code are out
of sync. At some point, it may be reasonable for pdserver to automatically
bootload the right software to the purpledrop.