Support feedback control (#7)
Add support for latest embedded software (v0.4.x) with updated messaging for feedback control.

This does not maintain backwards compatibility with older revisions. To prevent confusion, an embedded software version check is added to print a warning message to the user if an expected version is not found.

Also adjusts warnings on message parsing to use logger for output and to add a warning when a message is aborted without finishing, as this is likely a sign of something gone wrong.